requirements.txt
alabaster==0.7.12
asgiref==3.5.2
Babel==2.10.3
black==22.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.1
django-countries==7.3.2
django-crispy-forms==1.14.0
django-filter==22.1
django-seed==0.3.1
docutils==0.19
factory-boy==3.2.1
Faker==15.0.0
flake8==5.0.4
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.2.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
toposort==1.7
typing_extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
